divide up imageutils
