+++ pkg/image/apiserver/registry/imagesignature/rest.go
@@ -1,6 +1,7 @@
 package imagesignature
 
 import (
+	"bytes"
 	"context"
 	"errors"
 
@@ -12,6 +13,7 @@ import (
 	imagegroup "github.com/openshift/api/image"
 	imageapi "github.com/openshift/origin/pkg/image/apis/image"
 	imageclient "github.com/openshift/origin/pkg/image/generated/internalclientset/typed/image/internalversion"
+	"github.com/openshift/origin/pkg/image/util"
 )
 
 // REST implements the RESTStorage interface for ImageSignature
@@ -51,7 +53,7 @@ func (r *REST) Create(ctx context.Context, obj runtime.Object, createValidation 
 		}
 	}
 
-	imageName, _, err := imageapi.SplitImageSignatureName(signature.Name)
+	imageName, _, err := util.SplitImageSignatureName(signature.Name)
 	if err != nil {
 		return nil, kapierrors.NewBadRequest(err.Error())
 	}
@@ -62,7 +64,7 @@ func (r *REST) Create(ctx context.Context, obj runtime.Object, createValidation 
 	}
 
 	// ensure that given signature already doesn't exist - either by its name or type:content
-	if byName, byContent := imageapi.IndexOfImageSignatureByName(image.Signatures, signature.Name), imageapi.IndexOfImageSignature(image.Signatures, signature.Type, signature.Content); byName >= 0 || byContent >= 0 {
+	if byName, byContent := indexOfImageSignatureByName(image.Signatures, signature.Name), indexOfImageSignature(image.Signatures, signature.Type, signature.Content); byName >= 0 || byContent >= 0 {
 		return nil, kapierrors.NewAlreadyExists(imagegroup.Resource("imageSignatures"), signature.Name)
 	}
 
@@ -73,7 +75,7 @@ func (r *REST) Create(ctx context.Context, obj runtime.Object, createValidation 
 		return nil, err
 	}
 
-	byName := imageapi.IndexOfImageSignatureByName(image.Signatures, signature.Name)
+	byName := indexOfImageSignatureByName(image.Signatures, signature.Name)
 	if byName < 0 {
 		return nil, kapierrors.NewInternalError(errors.New("failed to store given signature"))
 	}
@@ -82,7 +84,7 @@ func (r *REST) Create(ctx context.Context, obj runtime.Object, createValidation 
 }
 
 func (r *REST) Delete(ctx context.Context, name string, options *metav1.DeleteOptions) (runtime.Object, bool, error) {
-	imageName, _, err := imageapi.SplitImageSignatureName(name)
+	imageName, _, err := util.SplitImageSignatureName(name)
 	if err != nil {
 		return nil, false, kapierrors.NewBadRequest("ImageSignatures must be accessed with <imageName>@<signatureName>")
 	}
@@ -92,7 +94,7 @@ func (r *REST) Delete(ctx context.Context, name string, options *metav1.DeleteOp
 		return nil, false, err
 	}
 
-	index := imageapi.IndexOfImageSignatureByName(image.Signatures, name)
+	index := indexOfImageSignatureByName(image.Signatures, name)
 	if index < 0 {
 		return nil, false, kapierrors.NewNotFound(imagegroup.Resource("imageSignatures"), name)
 	}
@@ -106,4 +108,26 @@ func (r *REST) Delete(ctx context.Context, name string, options *metav1.DeleteOp
 	}
 
 	return &metav1.Status{Status: metav1.StatusSuccess}, true, nil
+}
+
+// IndexOfImageSignatureByName returns an index of signature identified by name in the image if present. It
+// returns -1 otherwise.
+func indexOfImageSignatureByName(signatures []imageapi.ImageSignature, name string) int {
+	for i := range signatures {
+		if signatures[i].Name == name {
+			return i
+		}
+	}
+	return -1
+}
+
+// IndexOfImageSignature returns index of signature identified by type and blob in the image if present. It
+// returns -1 otherwise.
+func indexOfImageSignature(signatures []imageapi.ImageSignature, sType string, sContent []byte) int {
+	for i := range signatures {
+		if signatures[i].Type == sType && bytes.Equal(signatures[i].Content, sContent) {
+			return i
+		}
+	}
+	return -1
 }
