+++ pkg/oc/cli/debug/debug.go
@@ -51,9 +51,9 @@ import (
 	"github.com/openshift/library-go/pkg/image/reference"
 	"github.com/openshift/oc/pkg/helpers/conditions"
 	utilenv "github.com/openshift/oc/pkg/helpers/env"
-	imageapi "github.com/openshift/origin/pkg/image/apis/image"
-	imageutilinternal "github.com/openshift/origin/pkg/image/util"
+	"github.com/openshift/origin/pkg/image/util"
 	generateapp "github.com/openshift/origin/pkg/oc/lib/newapp/app"
+	"github.com/openshift/origin/pkg/oc/lib/ocimageutil"
 )
 
 const (
@@ -505,7 +505,7 @@ func (o *DebugOptions) getContainerImageViaDeploymentConfig(pod *corev1.Pod, con
 			trigger.ImageChangeParams != nil &&
 			trigger.ImageChangeParams.From.Kind == "ImageStreamTag" {
 
-			isname, _, err := imageapi.ParseImageStreamTagName(trigger.ImageChangeParams.From.Name)
+			isname, _, err := util.ParseImageStreamTagName(trigger.ImageChangeParams.From.Name)
 			if err != nil {
 				return nil, err
 			}
@@ -577,7 +577,7 @@ func (o *DebugOptions) getContainerImageCommand(pod *corev1.Pod, container *core
 		return nil, fmt.Errorf("error: no usable image found")
 	}
 
-	if err := imageutilinternal.ImageWithMetadata(image); err != nil {
+	if err := ocimageutil.ImageWithMetadata(image); err != nil {
 		return nil, err
 	}
 	dockerImage, ok := image.DockerImageMetadata.Object.(*dockerv10.DockerImage)
