+++ pkg/oc/lib/ocimageutil/helpers.go
@@ -1,0 +1,70 @@
+package ocimageutil
+
+import (
+	"encoding/json"
+	"strings"
+
+	dockerv10 "github.com/openshift/api/image/docker10"
+	imagev1 "github.com/openshift/api/image/v1"
+	imageapi "github.com/openshift/origin/pkg/image/apis/image"
+	"github.com/openshift/origin/pkg/oc/lib/ocimageutil/internal/digest"
+)
+
+// ImageWithMetadata mutates the given image. It parses raw DockerImageManifest data stored in the image and
+// fills its DockerImageMetadata and other fields.
+// Copied from github.com/openshift/image-registry/pkg/origin-common/util/util.go
+func ImageWithMetadata(image *imagev1.Image) error {
+	// Check if the metadata are already filled in for this image.
+	meta, hasMetadata := image.DockerImageMetadata.Object.(*dockerv10.DockerImage)
+	if hasMetadata && meta.Size > 0 {
+		return nil
+	}
+
+	version := image.DockerImageMetadataVersion
+	if len(version) == 0 {
+		version = "1.0"
+	}
+
+	obj := &dockerv10.DockerImage{}
+	if len(image.DockerImageMetadata.Raw) != 0 {
+		if err := json.Unmarshal(image.DockerImageMetadata.Raw, obj); err != nil {
+			return err
+		}
+		image.DockerImageMetadata.Object = obj
+	}
+
+	image.DockerImageMetadataVersion = version
+
+	return nil
+}
+
+// StatusHasTag returns named tag from image stream's status and boolean whether one was found.
+func StatusHasTag(stream *imagev1.ImageStream, name string) (imagev1.NamedTagEventList, bool) {
+	for _, tag := range stream.Status.Tags {
+		if tag.Tag == name {
+			return tag, true
+		}
+	}
+	return imagev1.NamedTagEventList{}, false
+}
+
+func HasAnnotationTag(tagRef *imagev1.TagReference, searchTag string) bool {
+	for _, tag := range strings.Split(tagRef.Annotations["tags"], ",") {
+		if tag == searchTag {
+			return true
+		}
+	}
+	return false
+}
+
+// ShortDockerImageID returns a short form of the provided DockerImage ID for display
+func ShortDockerImageID(image *imageapi.DockerImage, length int) string {
+	id := image.ID
+	if s, err := digest.ParseDigest(id); err == nil {
+		id = s.Hex()
+	}
+	if len(id) > length {
+		id = id[:length]
+	}
+	return id
+}
