+++ pkg/quota/image/imagestreamtag_evaluator.go
@@ -3,6 +3,8 @@ package image
 import (
 	"fmt"
 
+	"github.com/openshift/origin/pkg/image/util"
+
 	corev1 "k8s.io/api/core/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -92,7 +94,7 @@ func (i *imageStreamTagEvaluator) Usage(item runtime.Object) (corev1.ResourceLis
 		imagev1.ResourceImageStreams: *resource.NewQuantity(0, resource.BinarySI),
 	}
 
-	isName, _, err := imageapi.ParseImageStreamTagName(ist.Name)
+	isName, _, err := util.ParseImageStreamTagName(ist.Name)
 	if err != nil {
 		return corev1.ResourceList{}, err
 	}
